requirements.txt
customtkinter==5.2.2
mysql-connector-python==9.0.0
Pillow==11.0.0
darkdetect==0.8.0